requirements.txt
numpy>=1.24.0,<1.26.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
chardet>=5.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml
